Separate Session related functionality from Connection class (#571)
* decouple session class from existing Connection

ensure maintenance of current APIs of Connection while delegating
responsibility


* add open property to Connection to ensure maintenance of existing API


* update unit tests to address ThriftBackend through session instead of through Connection


* chore: move session specific tests from test_client to test_session


* formatting (black)

as in CONTRIBUTING.md


* use connection open property instead of long chain through session


* trigger integration workflow


* fix: ensure open attribute of Connection never fails

in case the openSession takes long, the initialisation of the session
will not complete immediately. This could make the session attribute
inaccessible. If the Connection is deleted in this time, the open()
check will throw because the session attribute does not exist. Thus, we
default to the Connection being closed in this case. This was not an
issue before because open was a direct attribute of the Connection
class. Caught in the integration tests.


* fix: de-complicate earlier connection open logic

earlier, one of the integration tests was failing because 'session was
not an attribute of Connection'. This is likely tied to a local
configuration issue related to unittest that was causing an error in the
test suite itself. The tests are now passing without checking for the
session attribute.
https://github.com/databricks/databricks-sql-python/pull/567/commits/c676f9b0281cc3e4fe9c6d8216cc62fc75eade3b


* Revert "fix: de-complicate earlier connection open logic"

This reverts commit d6b1b196c98a6e9d8e593a88c34bbde010519ef4.


* [empty commit] attempt to trigger ci e2e workflow


* Update CODEOWNERS (#562)

new codeowners


* Enhance Cursor close handling and context manager exception management to prevent server side resource leaks (#554)

* Enhance Cursor close handling and context manager exception management

* tests

* fmt

* Fix Cursor.close() to properly handle CursorAlreadyClosedError

* Remove specific test message from Cursor.close() error handling

* Improve error handling in connection and cursor context managers to ensure proper closure during exceptions, including KeyboardInterrupt. Add tests for nested cursor management and verify operation closure on server-side errors.

* add

* add


* PECOBLR-86 improve logging on python driver (#556)

* PECOBLR-86 Improve logging for debug level


* PECOBLR-86 Improve logging for debug level


* fixed format


* used lazy logging


* changed debug to error logs


* used lazy logging


---------


* Revert "Merge remote-tracking branch 'upstream/sea-migration' into decouple-session"

This reverts commit dbb2ec52306b91072a2ee842270c7113aece9aff, reversing
changes made to 7192f117279d4f0adcbafcdf2238c18663324515.


* Reapply "Merge remote-tracking branch 'upstream/sea-migration' into decouple-session"

This reverts commit bdb83817f49e1d88a01679b11da8e55e8e80b42f.


* fix: separate session opening logic from instantiation

ensures correctness of self.session.open call in Connection


* fix: use is_open attribute to denote session availability


* fix: access thrift backend through session


* chore: use get_handle() instead of private session attribute in client


* formatting (black)


* fix: remove accidentally removed assertions


---------
